+++ src/app/app-routing.module.ts
@@ -3,7 +3,6 @@ import {Routes, RouterModule} from "@angular/router";
 
 import {HomePageComponent} from "./HomePage/homePage.component";
 import {HelpPageComponent} from "./HelpPage/helpPage.component";
-import {ThirdPartyLicensesPageComponent} from "./ThirdPartyLicensesPage/thirdPartyLicensesPage.component";
 import {ServerPageComponent} from "./server-page/server-page.component";
 import {SettingsPageComponent} from "./settings-page/settings-page.component";
 
@@ -15,10 +14,6 @@ const routes: Routes = [
   {
     path: "help",
     component: HelpPageComponent
-  },
-  {
-    path: "3rdpartylicenses",
-    component: ThirdPartyLicensesPageComponent
   },
   {
     path: "s/:address/:port",